add validator tests
